feat: substitui Rank Math por Yoast SEO, atualiza Site Kit

+++ plugins/google-site-kit/includes/Core/Authentication/Authentication.php
@@ -156,14 +156,6 @@ final class Authentication implements Provides_Feature_Metrics {
 	protected $token;
 
 	/**
-	 * Owner_ID instance.
-	 *
-	 * @since 1.16.0
-	 * @var Owner_ID
-	 */
-	protected $owner_id;
-
-	/**
 	 * Has_Connected_Admins instance.
 	 *
 	 * @since 1.14.0
@@ -257,7 +249,6 @@ final class Authentication implements Provides_Feature_Metrics {
 		$this->verification_file              = new Verification_File( $this->user_options );
 		$this->profile                        = new Profile( $this->user_options );
 		$this->token                          = new Token( $this->user_options );
-		$this->owner_id                       = new Owner_ID( $this->options );
 		$this->has_connected_admins           = new Has_Connected_Admins( $this->options, $this->user_options );
 		$this->has_multiple_admins            = new Has_Multiple_Admins( $this->transients );
 		$this->connected_proxy_url            = new Connected_Proxy_URL( $this->options );
@@ -277,7 +268,7 @@ final class Authentication implements Provides_Feature_Metrics {
 		$this->verification_file()->register();
 		$this->verification_meta()->register();
 		$this->has_connected_admins->register();
-		$this->owner_id->register();
+		$this->has_multiple_admins->register();
 		$this->connected_proxy_url->register();
 		$this->disconnected_reason->register();
 		$this->initial_version->register();
@@ -311,9 +302,7 @@ final class Authentication implements Provides_Feature_Metrics {
 				}
 
 				$this->set_connected_proxy_url();
-			},
-			10,
-			3
+			}
 		);
 
 		add_filter(
@@ -330,7 +319,6 @@ final class Authentication implements Provides_Feature_Metrics {
 				}
 
 				$user['connectURL']           = esc_url_raw( $this->get_connect_url() );
-				$user['hasMultipleAdmins']    = $this->has_multiple_admins->get();
 				$user['initialVersion']       = $this->initial_version->get();
 				$user['isUserInputCompleted'] = ! $this->user_input->are_settings_empty();
 				$user['verified']             = $this->verification->has();
@@ -842,7 +830,6 @@ final class Authentication implements Provides_Feature_Metrics {
 	 * @return array Filtered $data.
 	 */
 	private function inline_js_base_data( $data ) {
-		$data['isOwner']             = $this->owner_id->get() === get_current_user_id();
 		$data['splashURL']           = esc_url_raw( $this->context->admin_url( 'splash' ) );
 		$data['proxySetupURL']       = '';
 		$data['proxyPermissionsURL'] = '';
@@ -1144,24 +1131,57 @@ final class Authentication implements Provides_Feature_Metrics {
 						?>
 						<a
 							href="#"
-							onclick="clearSiteKitAppStorage()"
+							onclick="reauthenticateAndContinueSetup()"
 						><?php esc_html_e( 'Click here', 'google-site-kit' ); ?></a>
 					</p>
 					<?php
 					BC_Functions::wp_print_inline_script_tag(
 						sprintf(
 							"
-							function clearSiteKitAppStorage() {
-								if ( localStorage ) {
-									localStorage.clear();
+							function reauthenticateAndContinueSetup() {
+								const moduleSlug = getAbandonedModuleSlug();
+
+								if ( moduleSlug ) {
+									const redirect = '%3\$s&slug=' + moduleSlug;
+									document.location = '%2\$s&redirect=' + encodeURIComponent( redirect );
+								} else {
+									if ( localStorage ) {
+										localStorage.clear();
+									}
+									if ( sessionStorage ) {
+										sessionStorage.clear();
+									}
+									document.location = '%2\$s';
 								}
-								if ( sessionStorage ) {
-									sessionStorage.clear();
+							}
+
+							function getAbandonedModuleSlug() {
+								for ( const storage of [ localStorage, sessionStorage ] ) {
+									if ( ! storage ) {
+										continue;
+									}
+									const key = Object.keys( storage ).find( ( k ) =>
+										k.match( 'googlesitekit_%1\$s_.*_module_setup' )
+									);
+									if ( ! key ) {
+										continue;
+									}
+									try {
+										return JSON.parse( storage[ key ] )?.value;
+									} catch ( _ ) {}
 								}
-								document.location = '%s';
 							}
 							",
-							esc_url_raw( $this->get_connect_url() )
+							GOOGLESITEKIT_VERSION,
+							esc_url_raw( $this->get_connect_url() ),
+							esc_url_raw(
+								$this->context->admin_url(
+									'dashboard',
+									array(
+										'reAuth' => 'true',
+									)
+								)
+							)
 						)
 					);
 					return ob_get_clean();
@@ -1355,17 +1375,6 @@ final class Authentication implements Provides_Feature_Metrics {
 	 */
 	public function get_has_multiple_admins_instance() {
 		return $this->has_multiple_admins;
-	}
-
-	/**
-	 * Helper method to return owner_id property.
-	 *
-	 * @since 1.131.0
-	 *
-	 * @return Owner_ID
-	 */
-	public function get_owner_id_instance() {
-		return $this->owner_id;
 	}
 
 	/**
